% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_adjusted_counts_for_unwanted_variation_sc}
\alias{add_adjusted_counts_for_unwanted_variation_sc}
\title{Add addjust counts for unwanted variation}
\usage{
add_adjusted_counts_for_unwanted_variation_sc(.data, .formula,
  do.scale = F, do.center = F)
}
\arguments{
\item{.data}{A seurat list}

\item{.formula}{A formula}

\item{do.scale}{A boolean}

\item{do.center}{A boolean}
}
\value{
A tt seurat object
}
\description{
Add addjust counts for unwanted variation
}
