% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_assay}
\alias{add_assay}
\title{Add assay to a seurat object}
\usage{
add_assay(var, assay, assay_name)
}
\arguments{
\item{var}{A seurat object}

\item{assay}{A seurat assay}

\item{assay_name}{An character name of the assay}
}
\value{
A seurat object
}
\description{
Add assay to a seurat object
}
