% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_cell_cycle_annotation_sc}
\alias{add_cell_cycle_annotation_sc}
\title{Add cell cyle annotation}
\usage{
add_cell_cycle_annotation_sc(input.df, .cell)
}
\arguments{
\item{input.df}{A tt object}
}
\value{
A tt object
}
\description{
Add cell cyle annotation
}
