% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_cluster_annotation_SNN_sc}
\alias{add_cluster_annotation_SNN_sc}
\title{Add cluster information of single cells}
\usage{
add_cluster_annotation_SNN_sc(.data, ...)
}
\arguments{
\item{.data}{A tt object}
}
\value{
A tt tt object
}
\description{
Add cluster information of single cells
}
