% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_doublet_classification_sc}
\alias{add_doublet_classification_sc}
\title{Add doublet classification}
\usage{
add_doublet_classification_sc(.data, doublet.reps)
}
\arguments{
\item{.data}{A tt object}

\item{doublet.reps}{A numeric}
}
\value{
A tt object
}
\description{
Add doublet classification
}
