% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_mitochndrion_transcription_abundance_sc}
\alias{add_mitochndrion_transcription_abundance_sc}
\title{Add mitochondrion transcription abundance}
\usage{
add_mitochndrion_transcription_abundance_sc(.data, .cell)
}
\arguments{
\item{.data}{A tt object}
}
\value{
A tt object
}
\description{
Add mitochondrion transcription abundance
}
