% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_normalised_counts_sc}
\alias{add_normalised_counts_sc}
\title{Add normalised counts}
\usage{
add_normalised_counts_sc(input.df)
}
\arguments{
\item{input.df}{A tt object}
}
\value{
A tt object
}
\description{
Add normalised counts
}
