% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_reduced_dimensions_TSNE}
\alias{add_reduced_dimensions_TSNE}
\title{Add TSNE dimensions}
\usage{
add_reduced_dimensions_TSNE(input.df, components = 10)
}
\arguments{
\item{input.df}{A tt object}

\item{components}{An integer, the number of components to calculate}
}
\value{
A tt object
}
\description{
Add TSNE dimensions
}
