% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_reduced_dimensions_UMAP}
\alias{add_reduced_dimensions_UMAP}
\title{Add UMAP dimensions}
\usage{
add_reduced_dimensions_UMAP(input.df, components = 10)
}
\arguments{
\item{input.df}{A tt object}

\item{components}{An integer, the number of components to calculate}
}
\value{
A tt object
}
\description{
Add UMAP dimensions
}
