% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_variable_genes_classification}
\alias{add_variable_genes_classification}
\title{Add variable gene annotation}
\usage{
add_variable_genes_classification(.data)
}
\arguments{
\item{.data}{A tt object}
}
\value{
A tt object
}
\description{
Add variable gene annotation
}
