% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{aggregate_cells}
\alias{aggregate_cells}
\title{Creates a `tt` object}
\usage{
aggregate_cells(
  .data,
  .sample = NULL,
  slot = "data",
  assays = NULL,
  aggregation_function = Matrix::rowSums
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}
}
\value{
A `tbl_df` object
}
\description{
aggregate_cells() creates a speudo-bulk `tbl` object formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |
}
\details{
\lifecycle{maturing}

...
}
\examples{




my_tt =  aggregate_cells(aggregate_cells::counts_mini, sample, transcript, count)


}
