% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate_cells,SingleCellExperiment-method}
\alias{aggregate_cells,SingleCellExperiment-method}
\title{aggregate_cells}
\usage{
\S4method{aggregate_cells}{SingleCellExperiment}(
  .data,
  .sample = NULL,
  slot = "data",
  aggregation_function = Matrix::rowSums
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}
}
\value{
A `aggregate_cells` object
}
\description{
aggregate_cells
}
