% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate_cells,tidyseurat-method}
\alias{aggregate_cells,tidyseurat-method}
\title{aggregate_cells}
\usage{
\S4method{aggregate_cells}{tidyseurat}(.data)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}
}
\value{
A `aggregate_cells` object
}
\description{
aggregate_cells
}
