% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{create_seurat_from_tibble_wide}
\alias{create_seurat_from_tibble_wide}
\title{Create seurat object from tibble}
\usage{
create_seurat_from_tibble_wide(.data, .sample, transcript_col_names, .cell,
  min.transcripts = 400, min.cells = 5, ...)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{A symbol for the sample column}

\item{.cell}{A symbol for the cell column}

\item{.transcript}{A symbol for the transcript name column}

\item{.abundance}{A symbol for counts}
}
\value{
A tt object
}
\description{
Create seurat object from tibble
}
