% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{create_tt_from_cellRanger_sc}
\alias{create_tt_from_cellRanger_sc}
\title{Create tt object from cellRanger results}
\usage{
create_tt_from_cellRanger_sc(
  dir_names,
  min.transcripts = 400,
  min.cells = 5,
  high.mito.thresh = 0.08,
  high.umi.thresh = 10000,
  species,
  genome = ifelse(species == "Human", "hg38", "mm10")
)
}
\arguments{
\item{dir_names}{A character array with the output directories of cellRanger if run from internal function}

\item{min.transcripts}{An integer with the threshold of the minimum number of genes allowed for each cell}

\item{min.cells}{An integer with the threshold of the minimum anumber of cells allowed for each sample}

\item{high.mito.thresh}{An numeric with the threshold of the maximum fraction of reads from mitochondrion. Used for filtering}

\item{high.umi.thresh}{An numeric with the threshold of the maximum fraction of umis. Used for filtering}

\item{species}{A character name of the species}

\item{genome}{A character name of the mapping genome used}
}
\value{
A tt object
}
\description{
Create tt object from cellRanger results
}
