% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{create_tt_from_seurat}
\alias{create_tt_from_seurat}
\title{Create tt object from seurat object}
\usage{
create_tt_from_seurat(seurat_object, dir_names = list(""),
  min.transcripts = 400, min.cells = 5, high.mito.thresh = 0.08,
  high.umi.thresh = 10000, sample_column = sample,
  cell_column = cell, species, genome = ifelse(species == "Human",
  "hg38", "mm10"))
}
\arguments{
\item{seurat_object}{A seurat object}

\item{dir_names}{A character array with the output directories of cellRanger if run from internal function}

\item{min.transcripts}{An integer with the threshold of the minimum number of genes allowed for each cell}

\item{min.cells}{An integer with the threshold of the minimum anumber of cells allowed for each sample}

\item{high.mito.thresh}{An numeric with the threshold of the maximum fraction of reads from mitochondrion. Used for filtering}

\item{high.umi.thresh}{An numeric with the threshold of the maximum fraction of umis. Used for filtering}

\item{sample_column}{A symbol for the sample column}

\item{cell_column}{A symbol for the cell column}

\item{species}{A character name of the species}

\item{genome}{A character name of the mapping genome used}
}
\value{
A tt object
}
\description{
Create tt object from seurat object
}
