% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{create_tt_from_tibble_wide_sc}
\alias{create_tt_from_tibble_wide_sc}
\title{Create tt object from tibble}
\usage{
create_tt_from_tibble_wide_sc(.data, .sample, .transcript_position, .cell,
  species, min.transcripts = 400, min.cells = 5,
  high.mito.thresh = 0.08, high.umi.thresh = 10000,
  genome = ifelse(species == "Human", "hg38", "mm10"), ...)
}
\arguments{
\item{.data}{A tibble}

\item{.sample}{A character name of the sample column}

\item{.cell}{A character name of the cell name column}

\item{species}{A character name of the species used (e.g., Mouse, Human)}

\item{.transcript}{A character name of the transcript name column}

\item{.abundance}{A character name of the count column}
}
\value{
A tibble with an additional column
}
\description{
Create tt object from tibble
}
