% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{do_integration_seurat}
\alias{do_integration_seurat}
\title{Integrate many seurat objects}
\usage{
do_integration_seurat(seurat_list, reference = NULL)
}
\arguments{
\item{.data}{A seurat list}
}
\value{
A tt seurat object
}
\description{
Integrate many seurat objects
}
