% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{drop_assay}
\alias{drop_assay}
\title{Drop assay of a seurat object}
\usage{
drop_assay(var, assay_name)
}
\arguments{
\item{var}{A seurat object}

\item{assay_name}{An character name of the assay}
}
\value{
A seurat object
}
\description{
Drop assay of a seurat object
}
