% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_adjusted_counts_for_unwanted_variation_sc}
\alias{get_adjusted_counts_for_unwanted_variation_sc}
\title{Get adjusted counts for unwanted variation}
\usage{
get_adjusted_counts_for_unwanted_variation_sc(.data, .formula,
  do.scale = F, do.center = F, verbose = T, ...)
}
\arguments{
\item{.data}{A seurat list}

\item{.formula}{A formula}

\item{do.scale}{A boolean}

\item{do.center}{A boolean}
}
\value{
A tt seurat object
}
\description{
Get adjusted counts for unwanted variation
}
