% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_cell}
\alias{get_cell}
\title{Get column names either from user or from attributes}
\usage{
get_cell(input.df, .cell)
}
\arguments{
\item{input.df}{A tibble}

\item{elements_column}{A character name of the sample column}
}
\value{
A list of column enquo or error
}
\description{
Get column names either from user or from attributes
}
