% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_cell_cycle_annotation_sc}
\alias{get_cell_cycle_annotation_sc}
\title{Get cell cyle annotation}
\usage{
get_cell_cycle_annotation_sc(input.df, .cell)
}
\arguments{
\item{input.df}{A tt object}
}
\value{
A tt object
}
\description{
Get cell cyle annotation
}
