% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_cell_type_annotation_sc}
\alias{get_cell_type_annotation_sc}
\title{Get cell type information on single cells}
\usage{
get_cell_type_annotation_sc(.data)
}
\arguments{
\item{.data}{A tt object}
}
\value{
A tt tt object
}
\description{
Get cell type information on single cells
}
