% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_cluster_annotation_SNN_automatic}
\alias{get_cluster_annotation_SNN_automatic}
\title{Get cluster information of single cells}
\usage{
get_cluster_annotation_SNN_automatic(.data, ...)
}
\arguments{
\item{.data}{A tt object}
}
\value{
A tt tt object
}
\description{
Get cluster information of single cells
}
