% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_cluster_annotation_SNN_sc}
\alias{get_cluster_annotation_SNN_sc}
\title{Get cluster information of single cells}
\usage{
get_cluster_annotation_SNN_sc(input.df, ...)
}
\arguments{
\item{input.df}{A tt object}
}
\value{
A tt tt object
}
\description{
Get cluster information of single cells
}
