% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_mitochndrion_transcription_abundance_sc}
\alias{get_mitochndrion_transcription_abundance_sc}
\title{Calculate mitochondrion transcription abundance}
\usage{
get_mitochndrion_transcription_abundance_sc(input.df, cell_column)
}
\arguments{
\item{input.df}{A tt object}
}
\value{
A tt object
}
\description{
Calculate mitochondrion transcription abundance
}
