% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_normalised_counts_sc}
\alias{get_normalised_counts_sc}
\title{Get normalised counts}
\usage{
get_normalised_counts_sc(.data)
}
\arguments{
\item{.data}{A tt object}
}
\value{
A tt object
}
\description{
Get normalised counts
}
