% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_reduced_dimensions_PCA}
\alias{get_reduced_dimensions_PCA}
\title{Get principal components}
\usage{
get_reduced_dimensions_PCA(input.df, components = 10)
}
\arguments{
\item{input.df}{A tt object}

\item{components}{An integer, the number of components to calculate}
}
\value{
A tt object
}
\description{
Get principal components
}
