% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifelse3_pipe}
\alias{ifelse3_pipe}
\title{This is a generalisation of ifelse that acceots an object and return an objects}
\usage{
ifelse3_pipe(.x, .p1, .p2, .p3, .f1, .f2, .f3, .f4 = NULL)
}
\arguments{
\item{.x}{A tibble}

\item{.p1}{A boolean}

\item{.p2}{ELSE IF condition}

\item{.f1}{A function}

\item{.f2}{A function}

\item{.f3}{A function}
}
\value{
A tibble
}
\description{
This is a generalisation of ifelse that acceots an object and return an objects
}
