% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{iterate_cell_cycle_scoreing}
\alias{iterate_cell_cycle_scoreing}
\title{Add variable gene annotation
This function is needed for a bug in Seurat about binning
https://github.com/satijalab/seurat/issues/1227}
\usage{
iterate_cell_cycle_scoreing(x, s.features, g2m.features, StartNBin = 24)
}
\arguments{
\item{input.df}{A seurat object}
}
\value{
A seurat object
}
\description{
Add variable gene annotation
This function is needed for a bug in Seurat about binning
https://github.com/satijalab/seurat/issues/1227
}
