% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{merged_tt_object}
\alias{merged_tt_object}
\title{Add cell type information on single cells}
\usage{
merged_tt_object(...)
}
\arguments{
\item{input.df}{A tt object}
}
\value{
A tt tt object
}
\description{
Add cell type information on single cells
}
