% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rename_sample_if_samples_not_set_by_user}
\alias{rename_sample_if_samples_not_set_by_user}
\title{Check if sample already set by the user, otherwise take sample information from Seurat object
I should check it with parameters attribute if .sample is present
For the moment is fine like this}
\usage{
rename_sample_if_samples_not_set_by_user(input.df)
}
\arguments{
\item{input.df}{A seurat object}
}
\value{
A seurat object
}
\description{
Check if sample already set by the user, otherwise take sample information from Seurat object
I should check it with parameters attribute if .sample is present
For the moment is fine like this
}
