% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_default_assay_seurat}
\alias{set_default_assay_seurat}
\title{Set defauly assay of a seurat object}
\usage{
set_default_assay_seurat(var, assay_name)
}
\arguments{
\item{var}{A seurat object}

\item{assay_name}{An character name of the assay}
}
\value{
A seurat object
}
\description{
Set defauly assay of a seurat object
}
