% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidysc}
\alias{tidysc}
\alias{tidysc_long}
\alias{tidysc_wide}
\alias{tidysc_cell_ranger}
\title{Creates a `tt` object from a `tbl``}
\usage{
tidysc_long(
  .data,
  .sample,
  .cell,
  .transcript,
  .abundance,
  species,
  min.transcripts = 400,
  min.cells = 5,
  ...
)

tidysc_wide(
  .data,
  .sample,
  .cell,
  .transcript_position,
  species,
  min.transcripts = 400,
  min.cells = 5,
  ...
)

tidysc_cell_ranger(
  dir_names,
  min.transcripts = 400,
  min.cells = 5,
  high.mito.thresh = 0.08,
  high.umi.thresh = 10000,
  species,
  genome = ifelse(species == "Human", "hg38", "mm10")
)
}
\arguments{
\item{.data}{A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |}

\item{.sample}{The name of the sample column}

\item{.cell}{The name of the cell column}

\item{.transcript}{The name of the transcript/gene column}

\item{.abundance}{The name of the transcript/gene abundance column}
}
\value{
A `tidysc` object

A `tidysc` object

A `tidysc` object
}
\description{
tidysc_long() creates a `tt` object from a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

tidysc_wide() creates a `tt` object from a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

tidysc_cell_ranger() creates a `tt` object from a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |
}
\details{
\lifecycle{experimental}

This function created a tidysc object and is useful if you want
to avoid to specify .sample, .transcript and .abundance arguments all the times.
The tidysc object have an attribute called parameters where these three
arguments are stored as metadata. They can be extracted as attr(<object>, "parameters").

\lifecycle{experimental}

This function created a tidysc object and is useful if you want
to avoid to specify .sample, .transcript and .abundance arguments all the times.
The tidysc object have an attribute called parameters where these three
arguments are stored as metadata. They can be extracted as attr(<object>, "parameters").

\lifecycle{experimental}

This function created a tidysc object and is useful if you want
to avoid to specify .sample, .transcript and .abundance arguments all the times.
The tidysc object have an attribute called parameters where these three
arguments are stored as metadata. They can be extracted as attr(<object>, "parameters").
}
\examples{
\donttest{



my_tt =
    tidysc::counts \%>\%
    tidysc(sample, transcript, counts)

class(my_tt)

}

\donttest{



my_tt =
    tidysc::counts \%>\%
    tidysc(sample, transcript, counts)

class(my_tt)

}

\donttest{



my_tt =
    tidysc::counts \%>\%
    tidysc(sample, transcript, counts)

class(my_tt)

}

}
