% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ttSc_cell_ranger.default}
\alias{ttSc_cell_ranger.default}
\title{#' @export
ttSc_cell_ranger.default <- function(dir_names,
                                     min.transcripts = 400,
                                     min.cells = 5,
                                     high.mito.thresh = 0.08,
                                     high.umi.thresh = 10000,
                                     species,
                                     genome = ifelse(species == "Human", "hg38", "mm10"))
{
  print("This function cannot be applied to this object")
}}
\usage{
\method{ttSc_cell_ranger}{default}(
  dir_names,
  min.transcripts = 400,
  min.cells = 5,
  high.mito.thresh = 0.08,
  high.umi.thresh = 10000,
  species,
  genome = ifelse(species == "Human", "hg38", "mm10")
)
}
\description{
#' @export
ttSc_cell_ranger.default <- function(dir_names,
                                     min.transcripts = 400,
                                     min.cells = 5,
                                     high.mito.thresh = 0.08,
                                     high.umi.thresh = 10000,
                                     species,
                                     genome = ifelse(species == "Human", "hg38", "mm10"))
{
  print("This function cannot be applied to this object")
}
}
