% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update_object_sc}
\alias{update_object_sc}
\title{Update seurat attribute based on cell content of the data tt frame}
\usage{
update_object_sc(.data, .cell = NULL)
}
\arguments{
\item{var}{A tt object}

\item{assay}{A seurat assay}

\item{assay_name}{An character name of the assay}
}
\value{
A tt object
}
\description{
Update seurat attribute based on cell content of the data tt frame
}
