% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{add_cell_type_annotation_sc}
\alias{add_cell_type_annotation_sc}
\title{Add cell type information on single cells}
\usage{
add_cell_type_annotation_sc(.data)
}
\arguments{
\item{.data}{A tt object}
}
\value{
A tt tt object
}
\description{
Add cell type information on single cells
}
