% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sc.R
\name{get_reduced_dimensions_TSNE}
\alias{get_reduced_dimensions_TSNE}
\title{Get TSNE dimensions}
\usage{
get_reduced_dimensions_TSNE(.data, .dims = 10)
}
\arguments{
\item{.data}{A tt object}

\item{.dims}{An integer, the number of .dims to calculate}
}
\value{
A tt object
}
\description{
Get TSNE dimensions
}
